routine independent interface

+++ bot_requests/views.py
@@ -3,9 +3,7 @@ from django.http import HttpResponse
 from django.http.response import JsonResponse
 from django.utils import timezone
 from django.core.files.storage import FileSystemStorage
-from django.views import generic
-from django.views.decorators.csrf import csrf_exempt
-from django.contrib.auth.models import User, Group
+from django.contrib.auth.models import User
 
 #rest
 from rest_framework.parsers import JSONParser 
@@ -19,9 +17,10 @@ from ratelimit.decorators import ratelimit
 import json
 
 #app
-from .serializers import CreateRiderRequestSerializer, ImportClassificationRequestSerializer
-from .models import CreateRiderRequest, BotRequest, ImportClassificationRequest
+from .serializers import *
+from .models import *
 from .run import run_bot
+from .dic import routine_to_model, routine_to_serializer
 
 RATELIMRQ='200000/h'
 
@@ -33,89 +32,61 @@ class DetailView(generics.ListCreateAPIView):
     model = BotRequest
     template_name = 'bot_requests/detail.html'
   
-@api_view(['DELETE'])       
-def create_rider_request_delete(request, pk):
-    if pk is not None:
-        rq =CreateRiderRequest.objects.get(pk=pk)
-        delete_rq(request,rq)
-
-@api_view(['DELETE'])       
-def import_classification_request_delete(request, pk):
-    if pk is not None:
-        rq = ImportClassificationRequest.objects.get(pk=pk)
-        delete_rq(request,rq)
-
-@api_view(['GET'])    
-def create_rider_request_list(request, userid):
-    if request.method == 'GET':
-        bot_requests = CreateRiderRequest.objects.filter(author=userid)
-        requests_serializer = CreateRiderRequestSerializer(bot_requests, many=True)
-        return JsonResponse(requests_serializer.data, safe=False)   
-    else:
-        return JsonResponse({'error':'no GET request'}, status=status.HTTP_400_BAD_REQUEST)   
-
-@api_view(['GET'])    
-def import_classification_request_list(request, userid):
-    if request.method == 'GET':
-        bot_requests = ImportClassificationRequest.objects.filter(author=userid)
-        requests_serializer =ImportClassificationRequestSerializer(bot_requests, many=True)
-        return JsonResponse(requests_serializer.data, safe=False)   
-    else:
-        return JsonResponse({'error':'no GET request'}, status=status.HTTP_400_BAD_REQUEST)   
-
-@api_view(['GET'])    
-@permission_classes((IsAdminUser,))
-def all_create_rider_request_list(request, userid):
-    if request.method == 'GET':
-            bot_requests = CreateRiderRequest.objects.all()
-            requests_serializer = CreateRiderRequestSerializer(bot_requests, many=True)
-            return JsonResponse(requests_serializer.data, safe=False)   
-    else:
-        return JsonResponse({'error':'no GET request'}, status=status.HTTP_400_BAD_REQUEST)   
-
-@api_view(['GET'])   
-@permission_classes((IsAdminUser,)) 
-def all_import_classification_request_list(request, userid):
-    if request.method == 'GET':
-        bot_requests = ImportClassificationRequest.objects.all()
-        requests_serializer =ImportClassificationRequestSerializer(bot_requests, many=True)
-        return JsonResponse(requests_serializer.data, safe=False)   
-    else:
-        return JsonResponse({'error':'no GET request'}, status=status.HTTP_400_BAD_REQUEST)   
-
-
 ##### POST RQ #####
+def serial_save(request_serializer, request, rq_data):
+    if request_serializer.is_valid():
+        rq=request_serializer.save()
+        return run_autocheck(rq_data, request, rq.id)
+    else:
+        print('serializer error: ')
+        print(request_serializer.errors)
+        return JsonResponse(request_serializer.errors, status=status.HTTP_400_BAD_REQUEST)   
+
 @ratelimit(key='ip', rate=RATELIMRQ)
-@api_view(['POST'])    
-def create_rider_request(request):
+@api_view(['POST'])  
+def create_rq(request,routine):
     if request.method == 'POST':
+        serializer=routine_to_serializer(routine)
         rq_data=JSONParser().parse(request)
         rq_data.update(entry_time=timezone.now() ) 
-        rq_data.update(routine="create_rider")
-        rq_data.update(item_id="Q1")
-        request_serializer =CreateRiderRequestSerializer(data=rq_data)
+        rq_data.update(routine=routine)
+        
+        ##supplementary info depending on the routine
+        if routine in ["create_rider", "race", "national_all_champs", "national_one_champ"]:
+            rq_data.update(item_id="Q1")
+        
+        request_serializer =serializer(data=rq_data)
         return serial_save(request_serializer, request, rq_data)
-
-@ratelimit(key='ip', rate=RATELIMRQ)
-@api_view(['POST','GET'])    
-def import_classification_request(request):
-    if request.method == 'POST' and request.FILES:
+            
+def create_file_rq(request,routine):  
+     if request.method == 'POST' and request.FILES:
         try:
             #upload file
+            serializer=routine_to_serializer(routine)
             uploaded_file=request.FILES['file']
             fs = FileSystemStorage()
             name = fs.save(uploaded_file.name, uploaded_file)
             rq_data= (json.loads(request.POST['botrequest'] ))
             rq_data.update(entry_time=timezone.now() ) 
-            rq_data.update(routine="import_classification")
+            rq_data.update(routine=routine)
             rq_data.update(result_file_name=name)
 
-            request_serializer =ImportClassificationRequestSerializer(data=rq_data)
+            request_serializer =serializer(data=rq_data)
             return serial_save(request_serializer, request, rq_data)
         except:
             print("plantage by file")  
-            return JsonResponse({'file':'failed'}, status=status.HTTP_400_BAD_REQUEST) 
- 
+            return JsonResponse({'file':'failed'}, status=status.HTTP_400_BAD_REQUEST)    
+
+def run_autocheck(rq_data, request, rq_id):
+    user=User.objects.get(pk=rq_data["author"])
+    if user.has_perm('bot_requests.can_run_requests'):
+        run_error=run_bot(rq_id,rq_data["routine"])
+        if run_error==0:
+            return JsonResponse({'ras':'ras'}, status=status.HTTP_200_OK) 
+        else:
+            return JsonResponse({'error':'bot run not successful'}, status=status.HTTP_417_EXPECTATION_FAILED)        
+    return JsonResponse({'ras':'ras'}, status=status.HTTP_201_CREATED) 
+   
 @api_view(['POST'])    
 @permission_classes((IsAdminUser,))
 def run(request):
@@ -136,11 +107,15 @@ def run(request):
     else:
         return JsonResponse({'error':'no POST request'}, status=status.HTTP_400_BAD_REQUEST)   
 
-##sub-functions
-def delete_rq(request,rq):
+
+@api_view(['DELETE'])   
+def delete_rq(request,pk,routine):
     try:
         if request.method == 'DELETE':
-            rq.delete()
+            table=routine_to_model(routine)
+            if pk is not None:
+                rq =table.objects.get(pk=pk)
+                rq.delete()
             return HttpResponse(status=status.HTTP_204_NO_CONTENT)
         else:
             return JsonResponse({"delete":"failed"},status=status.HTTP_400_BAD_REQUEST)
@@ -148,22 +123,28 @@ def delete_rq(request,rq):
         print("plantage")
         return JsonResponse({"delete":"failed"},status=status.HTTP_400_BAD_REQUEST)
 
-def serial_save(request_serializer, request, rq_data):
-    if request_serializer.is_valid():
-        rq=request_serializer.save()
-        return run_autocheck(rq_data, request, rq.id)
+@api_view(['GET'])    
+def get_request_list(request, userid, routine):
+    if request.method == 'GET':
+        table=routine_to_model(routine)
+        serializer=routine_to_serializer(routine)
+        bot_requests = table.objects.filter(author=userid)
+        requests_serializer = serializer(bot_requests, many=True)
+        return JsonResponse(requests_serializer.data, safe=False)   
     else:
-        print('serializer error: ')
-        print(request_serializer.errors)
-        return JsonResponse(request_serializer.errors, status=status.HTTP_400_BAD_REQUEST)   
+        return JsonResponse({'error':'no GET request'}, status=status.HTTP_400_BAD_REQUEST) 
 
-def run_autocheck(rq_data, request, rq_id):
-    user=User.objects.get(pk=rq_data["author"])
-    if user.has_perm('bot_requests.can_run_requests'):
-        run_error=run_bot(rq_id,rq_data["routine"])
-        if run_error==0:
-            return JsonResponse({'ras':'ras'}, status=status.HTTP_200_OK) 
-        else:
-            return JsonResponse({'error':'bot run not successful'}, status=status.HTTP_417_EXPECTATION_FAILED)        
-    return JsonResponse({'ras':'ras'}, status=status.HTTP_201_CREATED) 
+@api_view(['GET'])    
+@permission_classes((IsAdminUser,))
+def all_get_request_list(request, userid, routine):
+    if request.method == 'GET':
+            table=routine_to_model(routine)
+            serializer=routine_to_serializer(routine)
+            bot_requests = table.objects.all()
+            requests_serializer = serializer(bot_requests, many=True)
+            return JsonResponse(requests_serializer.data, safe=False)   
+    else:
+        return JsonResponse({'error':'no GET request'}, status=status.HTTP_400_BAD_REQUEST)
+    
+
